Fix two iOS audio issues
By commenting out AVAudioSessionCategoryOptionAllowBluetooth we only
disable bluetooth input, which is desirable, because the audio output
stream of a bidirectional bluetooth stream is 16KHz sample rate. By
not passing AVAudioSessionCategoryOptionAllowBluetooth, we get good
sample rates with bluetooth devices.

The AVAudioSessionCategoryOptionOverrideMutedMicrophoneInterruption
patch is related a privacy feature introduced in iOS 14.5, where the
microphone is muted when closing the iPad's Smart Folio (magnetic)
cover. If we don't pass
AVAudioSessionCategoryOptionOverrideMutedMicrophoneInterruption,
the audio stream will be interrupted when closing the cover. In
JUCE this is not handled correctly, and the stream is bidirectionally
interrupted and never recovers during the lifetime of the app. An
app restart is needed to get processBlock calls again.